Created new page for notes and button
