+++ src/components/Repos/Notes.json
@@ -1,0 +1,10 @@
+{
+  "initialNote": [
+    {
+      "note": "Good job!"
+    },
+    {
+      "note": "Nice code!"
+    }
+  ]
+}
