+++ src/components/Repos/index.js
@@ -1,4 +1,5 @@
 import Repos from "./Repos";
 import Repo from "./Repo";
+import AddNoteForm from "./AddNoteForm";
 
-export { Repos, Repo };
+export { Repos, Repo, AddNoteForm };
