--- src/db.json
@@ -1,11 +1,0 @@
-
-{
-  "initialNote" : [
-     {
-       "note" : "Good job!"
-     },
-     {
-       "note" : "Nice code!"
-     }
-   ]
- }
